+++ app/Http/Controllers/ProdutoController.php
@@ -14,6 +14,9 @@ class ProdutoController extends Controller
     }
 
     public function index(){
+        if(!$this->validar()){
+            return Redirect("/");
+        }
         $list_produtos = Produto::all();
         return view('produto.index', [
             'produtos' => $list_produtos
@@ -21,37 +24,53 @@ class ProdutoController extends Controller
     }
 
     public function create(){
+        if(!$this->validar()){
+            return Redirect("/");
+        }
         return view('produto.create');
     }
 
     public function salvar(Request $request){
+        if(!$this->validar()){
+            return Redirect("/");
+        }
         $produto = Produto::create($request->all());
         return redirect("/produto")->with("message", "Produto criada com sucesso!");
     }
 
     public function atualizar(Request $request){
+        if(!$this->validar()){
+            return Redirect("/");
+        }
         $produto = $this->getProduto($request->id);
         $produto->update($request->all());
         return redirect("/produto");
     }
 
     public function editarView($id) {
+        if(!$this->validar()){
+            return Redirect("/");
+        }
         return view('produto.edit', [
             'produto' => $this->getProduto($id)
         ]);
     }
 
     public function excluir($id) {
+        if(!$this->validar()){
+            return Redirect("/");
+        }
         $this->getProduto($id)->delete();
         return redirect(url('produto'))->with('success', 'Excluído!');
     }
     //relatórios
     public function relatorioProduto(){
-
-       $data = ['dado'=>'ola'];
+        if(!$this->validar()){
+            return Redirect("/");
+        }
+        $data = ['dado'=>'ola'];
         $pdf = PDF::loadView('produto.relatorio.vendas', $data);
         return $pdf->stream('produto.relatorio.vendas.pdf');
-       // return $pdf->dawnload('produto.relatorio.vendas.pdf')
     }
     protected function getProduto($id)  {
         return $this->produto->find($id);
